Fix server crash by try-catch around mongoose connection esstablish
Summary:
The server used to crash 'randomly' without useful error-logging.
This was because mongoose could not establish a database connection
via the provided url in `process.env.MONGO_PATH`. By fixing the url
when running in node docker container it is no fixed when run via
docker-compose.
The try-catch now also loggs a detailed error message.